% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_mod.R
\name{aug_mod}
\alias{aug_mod}
\title{Augment a paired competition model with new data}
\usage{
aug_mod(newdata, comp_mod, ...)
}
\arguments{
\item{newdata}{a data frame containing new data to be added to the original data used to build the model.}

\item{comp_mod}{a paired competition model object.}
}
\value{
an augmented data frame with the results of the paired competition model extended with the new data.
}
\description{
This function extends a paired competition model with new data, returning the augmented results.
}
\examples{
# Create a paired competition model comp_mod using your data and formula
# Define new data in the data frame newdata
augmented_model <- aug_mod(newdata, comp_mod)

}
