% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model_mat.R
\name{build_model_mat}
\alias{build_model_mat}
\title{Build a paired competition model matrix for generalized linear models (GLMs)}
\usage{
build_model_mat(
  formula,
  data,
  p1,
  p2,
  p1_effects = ~1,
  p2_effects = ~1,
  ref_player = NULL
)
}
\arguments{
\item{formula}{an object of class "formula". It should be a two-sided formula
with the numeric response variable on the left-hand side and contest-level
predictors on the right-hand side.}

\item{data}{a data frame with variables used to fit the model.}

\item{p1}{a string with the name of the column in the data frame that corresponds
to the first player.}

\item{p2}{a string with the name of the column in the data frame that corresponds
to the second player.}

\item{p1_effects}{a one-sided formula object specifying player-level effects for
the first player.}

\item{p2_effects}{a one-sided formula object specifying player-level effects for
the second player.}

\item{ref_player}{}
}
\value{
a list containing the model matrix components needed for competition modeling.
}
\description{
This function prepares the model matrix required for fitting competition models
}
\examples{
build_model_mat(point_winner ~ atp_importance, data = tennis_point,
      p1 = "player1", p2 = "player2",
      p1_effects = ~ point_server1, p2_effects = ~ point_server2,
      ref_player = "Milos Raonic")

build_model_mat(point_winner ~ -1, data = tennis_point,
    p1 = "player1", p2 = "player2",
    p1_effects = ~ 1, p2_effects = ~ 1,
    ref_player = "Milos Raonic")

build_model_mat(formula = point_winner ~ -1, data = tennis_point_wta,
    p1 = "player1", p2 = "player2",
    p1_effects = ~ 1, p2_effects = ~ 1,
    ref_player = "Angelique Kerber")

}
