\name{Cocktail_Cum}
\alias{Cocktail_Cum}
\title{
  Beverages paired comparison
}
\description{
    Paired comparison of 7 beverages by 112 subjects according their preferences
}
\usage{data(Cocktail)
}
\format{
  A matrix resulting of the cumulative paired comparison results of 7 products by 112 consumers. The (i,j) element correponds to the number of time product i was prefered to product j among all comparisons between these two products.
}

\examples{
  data(Cocktail_Cum)
  Cocktail_Cum
}