% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_glm.R
\name{comp_glm}
\alias{comp_glm}
\title{Fit a paired competition generalized linear model}
\usage{
comp_glm(
  formula,
  data,
  p1,
  p2,
  p1_effects = ~1,
  p2_effects = ~1,
  ref_player = NULL
)
}
\arguments{
\item{formula}{an object of class \code{"formula"}. \code{formula}
should be a two-sided formula with the numeric response variable on
 the left-hand side and contest-level predictors on the right-hand
 side.}

\item{data}{a data frame with variables to fit the model}

\item{p1}{a string with the name of the column of \code{data} with
the first player.}

\item{p2}{a string with the name of the column of \code{data} with
the second player.}

\item{p1_effects}{a one-sided \code{"formula"} object giving
player-level effects for the first player.}

\item{p2_effects}{a one-sided \code{"formula"} object giving
player-level effects for the second player.}

\item{ref_player}{a string giving the reference player}
}
\value{
a fitted model object of class \code{"glm"} and \code{"lm"}
}
\description{
Fit a competition model with either contest-level fixed effects
or player-level fixed effects.
}
\examples{
comp_glm(point_winner ~ atp_importance, data = tennis_point,
      p1 = "player1", p2 = "player2",
      p1_effects = ~ point_server1, p2_effects = ~ point_server2,
      ref_player = "Milos Raonic")

comp_glm(point_winner ~ -1, data = tennis_point,
    p1 = "player1", p2 = "player2",
    p1_effects = ~ 1, p2_effects = ~ 1,
    ref_player = "Milos Raonic")

comp_glm(formula = point_winner ~ -1, data = tennis_point_wta,
    p1 = "player1", p2 = "player2",
    p1_effects = ~ 1, p2_effects = ~ 1,
    ref_player = "Angelique Kerber")

comp_glm(formula = contest_winner ~ contest_quant + contest_fact,
   data = toy_comp,
   p1 = "player1", p2 = "player2",
   p1_effects = ~ player1_fact,
   p2_effects = ~ player2_fact)

}
