% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_vec.R
\name{create_d_mat}
\alias{create_d_mat}
\title{create a design matrix with 0's, 1's, and -1's.}
\usage{
create_d_mat(
  vec,
  players,
  player1_vec,
  player2_vec,
  val1 = 1,
  val2 = 1,
  names = NULL
)
}
\arguments{
\item{vec}{is a vector of 0's}

\item{players}{is a vector of strings (players) the same length as vec}

\item{player1_vec}{is a vector of "player" 1's}

\item{player2_vec}{is a vector of "player" 2's}

\item{val1}{is a value assigned to the index of \code{player1}}

\item{val2}{is a value assigned to the index of \code{player2}}

\item{names}{is an optional vector of names given to the columns of the
returned matrix.}
}
\value{
a matrix with the number of rows equal to the length of player1_vec and the number of columns equal to the length of players. Each row contains a single 1 (corresponding to player 1 for that row) and a single -1 (corresponding to player 2 for that row).
}
\description{
create a design matrix with 0's, 1's, and -1's.
}
\examples{
vec <- rep(0, 3)
players <- c("A", "B", "C")
player1 <- c("A", "B")
player2 <- c("C", "A")
create_d_mat(vec, players, player1, player2, val1 = c(1, 1), val2 = c(1, 1))
}
