% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a paired competition generalized linear model (GLM)}
\usage{
fit_model(mat_list, family)
}
\arguments{
\item{mat_list}{a list containing the model matrix components as generated
by the `build_model_mat` function.}

\item{family}{}

\item{...}{additional arguments to be passed to the glm function (e.g., family).}
}
\value{
a fitted model object of class "glm" and "lm" with competition modeling results.
comp_glm(point_winner ~ atp_importance, data = tennis_point,
      p1 = "player1", p2 = "player2",
      p1_effects = ~ point_server1, p2_effects = ~ point_server2,
      ref_player = "Milos Raonic", family = binomial)
}
\description{
This function fits a competition model using the model matrix components
prepared by `build_model_mat`. It allows you to specify additional arguments
}
