% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_vec.R
\name{get_data_vec}
\alias{get_data_vec}
\title{Replace values in a vector at certain indices with a 1 and -1}
\usage{
get_data_vec(vec, players, player1, player2, val1 = 1, val2 = 1)
}
\arguments{
\item{vec}{is a vector}

\item{players}{is a vector of strings (players) the same length as vec}

\item{player1}{is a string that gives the index that should take on a 1}

\item{player2}{is a string that gives the index that should take on a -1}

\item{val1}{is a value assigned to the index of \code{player1}}

\item{val2}{is a value assigned to the index of \code{player2}}
}
\value{
a vector with a 1 and -1 replacing two of the original values
}
\description{
Replace values in a vector at certain indices with a 1 and -1
}
\examples{
vec <- rep(0, 3)
players <- c("A", "B", "C")
get_data_vec(vec, players, "A", "C")
}
