\name{getCons}
\alias{getCons}
\title{
Gets the individuals labels.
}
\description{
Gets the individuals labels.
}
\usage{
getCons(object)
}
\arguments{
  \item{object}{
An object of class \code{DataPairComp}
}
}

\value{
vector of the individuals labels.
}
\examples{
data(Cocktail)
Cocktail_Cons<-getCons(Cocktail)
}
