\name{getCrit}
\alias{getCrit}
\title{
Gets the criteria's labels.}
\description{
Gets the criteria's labels.}
\usage{
getCrit(object)
}
\arguments{
  \item{object}{
An object of class \code{DataPairComp}}
}

\value{
vector of the criteria's labels.
}

\examples{
data(Cocktail)
Cocktail_Crit<-getCrit(Cocktail)
}