\name{getProd}
\alias{getProd}
\title{
Gets the products labels.}
\description{
Gets the products labels.
}
\usage{
getProd(object)
}
\arguments{
  \item{object}{
An object of class \code{DataPairComp}}
}
\value{
vector of the products labels.
}
\examples{
data(Cocktail)
Cocktail_Prod<-getProd(Cocktail)
}