\name{getTest}
\alias{getTest}
\title{
Gets the level and the quantile of Likelihood ratio test.
}
\description{
Gets the level and the quantile of Likelihood ratio test from the function \code{ResSimulLvrRatio()}
}
\usage{
getTest(object)
}
\arguments{
  \item{object}{
An object of class \code{LvrRatio}}
}
\value{
Matrix with the level and the associated quantile after performing Likelihood Ratio test.
}