% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba_game}
\alias{nba_game}
\title{NBA data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1230 rows and 7 columns.
}
\usage{
nba_game
}
\description{
A data set with NBA data.
}
\keyword{datasets}
