\name{ResCocktail1}
\alias{ResCocktail1}
\title{
  Result of \code{EstimBradley} function for 1 class and data \code{Cocktail}
}
\description{
    Result of \code{EstimBradley} function for 1 class and data \code{Cocktail}
}
\usage{data(ResCocktail1)
}
\format{
  A \code{BradleyEstim} class object with the following elements:
}

\examples{
data(ResCocktail1)
show(ResCocktail1)
}