\name{show-methods}
\docType{methods}
\alias{show-methods}
\title{ Methods for Function \code{show}}
\description{
Methods for function \code{show}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "BradleyEstim")}}{
}

\item{\code{signature(object = "DataPairComp")}}{
}
}}