% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis_match}
\alias{tennis_match}
\title{Tennis Match Data}
\format{
A data frame with 6 variables
\describe{
  \item{winner}{a \code{1} for all matches}
  \item{winner_name}{the name of the winning player}
  \item{loser_name}{the name of the losing player}
  \item{age_diff}{difference in player ages (winner minus loser)}
  \item{ht_diff}{difference in player heights (winner minus loser)}
  \item{cov}{a randomly generated covariate}
  ...
}
}
\usage{
tennis_match
}
\description{
A data set containing tennis matches in 2023 for 8 players.
}
\examples{
\dontrun{
players <- c("Carlos Alcaraz", "Daniil Medvedev", "Holger Rune",
"Novak Djokovic", "Jannik Sinner", "Taylor Fritz",
"Karen Khachanov", "Stefanos Tsitsipas")
tennis_match <- readr::read_csv("https://raw.githubusercontent.com/JeffSackmann/tennis_atp/master/atp_matches_2023.csv") |>
  dplyr::filter(winner_name \%in\% players &
                 loser_name \%in\% players) |>
 dplyr::select(winner_name, loser_name, winner_age, loser_age,
        winner_ht, loser_ht) |>
 dplyr::mutate(age_diff = winner_age - loser_age,
        ht_diff = winner_ht - loser_ht,
        winner = rep(1, length(winner_age)),
        cov = rnorm(length(winner_age), 0, 1)) |>
 dplyr::select(winner, winner_name, loser_name, age_diff, ht_diff, cov)
 save(tennis_match, file = "data/tennis_match.rda")
}
}
\keyword{datasets}
