% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis_point}
\alias{tennis_point}
\title{Tennis Point Data}
\format{
A data frame with 9 variables
\describe{
  \item{player1}{the name of the first player}
  \item{player2}{the name of the second player}
  \item{point_winner}{a \code{1} if the first player won the point and a \code{0} if the second player won the point}
  \item{atp_importance}{the importance of the point}
  \item{match_num}{a match identification number}
  \item{year}{the year the match was played in}
  \item{slam}{the slam (\code{wimbledon} for all of these matches)}
  \item{point_server1}{a \code{1} if the server was \code{player1}, and a \code{0} otherwise}
  \item{point_server2}{a \code{1} if the server was \code{player2}, and a \code{0} otherwise}
  ...
}
}
\usage{
tennis_point
}
\description{
A data set containing point-level data for men's singles matches
at Wimbledon from 2017 through 2021. Only matches from the 4th round
onward are included.
}
\examples{
\dontrun{
tennis_point <- readr::read_csv("inst/extra_data/test_data_serving.csv") |>
 dplyr::mutate(point_server1 = point_server,
        point_server2 = dplyr::if_else(point_server == 0,
                                true = 1, false = 0)) |>
 dplyr::select(-point_server)
save(tennis_point, file = "data/tennis_point.rda")
}
}
\keyword{datasets}
