% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_comp}
\alias{toy_comp}
\title{Toy Competition Data}
\format{
A data frame with 9 variables
\describe{
  \item{contest_winner}{a \code{1} if \code{player1} won the contest and a \code{0} if \code{player2} won the contest}
  \item{player1}{a vector of players}
  \item{player2}{a vector of competing players}
  \item{player1_quant}{a quantitative player-level covariate for \code{player1}}
  \item{player2_quant}{a quantitative player-level covariate for \code{player2}}
  \item{player1_fact}{a categorical player-level covariate for \code{player1}}
  \item{player2_fact}{a categorical player-level covariate for \code{player2}}
  \item{contest_fact}{a categorical contest-level covariate}
  \item{contest_quant}{a quantitative contest-level covariate}
  ...
}
}
\usage{
toy_comp
}
\description{
A data set with simulated competition data.
}
\examples{
\dontrun{
set.seed(05052023)
player1 <- sample(c("A", "B", "C", "D"), size = 150, replace = TRUE)
player2 <- sample(c("A", "B", "C", "D"), size = 150, replace = TRUE)
keep_ind <- player1 != player2
player1 <- player1[keep_ind]
player2 <- player2[keep_ind]
n <- length(player1)
contest_fact <- sample(c("Venue 1", "Venue 2", "Venue 3"), size = n,
                       replace = TRUE)
contest_quant <- runif(n, 0, 1)
player1_fact <- sample(c("W", "X", "Y", "Z"), size = n, replace = TRUE)
player2_fact <- sample(c("W", "X", "Y", "Z"), size = n, replace = TRUE)
player1_quant <- rnorm(n, 20, 2)
player2_quant <- rnorm(n, 20, 2)
contest_winner <- rbinom(n, size = 1, prob = 0.54)
toy_comp <- tibble::tibble(contest_winner, player1, player2,
                           player1_quant, player2_quant,
                           player1_fact, player2_fact,
                           contest_fact, contest_quant)
save(toy_comp, file = "data/toy_comp.rda")
}
}
\keyword{datasets}
